+++ test/cases/es-named-export-as-is/expected/main.js
@@ -12,9 +12,9 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony export */   cClass: () => (/* binding */ _3)
 /* harmony export */ });
 // extracted by mini-css-extract-plugin
-var _1 = "foo__style__a-class";
-var _2 = "foo__style__b__class";
-var _3 = "foo__style__cClass";
+var _1 = "Xh041yLR4iCP4RGjge50";
+var _2 = "NMuRsxoDwvW8BhSXhFAY";
+var _3 = "ayWIv09rPsAqE2JznIsI";
 
 
 
@@ -83,7 +83,7 @@ __webpack_require__.r(__webpack_exports__);
 
 
 // eslint-disable-next-line no-console
-console.log({ css: _style_css__WEBPACK_IMPORTED_MODULE_0__["default"], aClass: _style_css__WEBPACK_IMPORTED_MODULE_0__.aClass, bClass: _style_css__WEBPACK_IMPORTED_MODULE_0__.bClass, cClass: _style_css__WEBPACK_IMPORTED_MODULE_0__.cClass });
+console.log({ css: _style_css__WEBPACK_IMPORTED_MODULE_0__["default"], aClass: _style_css__WEBPACK_IMPORTED_MODULE_0__["a-class"], bClass: _style_css__WEBPACK_IMPORTED_MODULE_0__.b__class, cClass: _style_css__WEBPACK_IMPORTED_MODULE_0__.cClass });
 
 })();