+++ types/index.d.ts
@@ -13,6 +13,13 @@ declare class MiniCssExtractPlugin {
     webpack: Compiler["webpack"]
   ): CssDependencyConstructor;
   /**
+   * Returns all hooks for the given compilation
+   * @param {Compilation} compilation
+   */
+  static getCompilationHooks(
+    compilation: Compilation
+  ): MiniCssExtractPluginCompilationHooks;
+  /**
    * @param {PluginOptions} [options]
    */
   constructor(options?: PluginOptions | undefined);
@@ -94,6 +101,8 @@ declare namespace MiniCssExtractPlugin {
     CssDependency,
     CssDependencyOptions,
     CssDependencyConstructor,
+    VarNames,
+    MiniCssExtractPluginCompilationHooks,
   };
 }
 type Compiler = import("webpack").Compiler;
@@ -103,6 +112,13 @@ type CssDependencyConstructor = new (
   context: string | null,
   identifierIndex: number
 ) => CssDependency;
+type Compilation = import("webpack").Compilation;
+type MiniCssExtractPluginCompilationHooks = {
+  beforeTagInsert: import("tapable").SyncWaterfallHook<
+    [string, VarNames],
+    string
+  >;
+};
 type PluginOptions = {
   filename?: Required<Configuration>["output"]["filename"];
   chunkFilename?: Required<Configuration>["output"]["chunkFilename"];
@@ -166,7 +182,6 @@ declare const pluginName: "mini-css-extract-plugin";
 declare const pluginSymbol: unique symbol;
 declare var loader: string;
 type Schema = import("schema-utils/declarations/validate").Schema;
-type Compilation = import("webpack").Compilation;
 type ChunkGraph = import("webpack").ChunkGraph;
 type Chunk = import("webpack").Chunk;
 type ChunkGroup = Parameters<import("webpack").Chunk["isInGroup"]>[0];
@@ -233,3 +248,10 @@ type CssModuleDependency = {
 };
 type CssDependency = Dependency & CssModuleDependency;
 type CssDependencyOptions = Omit<LoaderDependency, "context">;
+type VarNames = {
+  tag: string;
+  chunkId: string;
+  href: string;
+  resolve: string;
+  reject: string;
+};